+++ hassio/src/hassio-pages-with-tabs.js
@@ -15,6 +15,9 @@ import './hassio-markdown-dialog.js';
 import './snapshots/hassio-snapshot.js';
 import './snapshots/hassio-snapshots.js';
 import './system/hassio-system.js';
+
+import scrollToTarget from '../../src/common/scroll-to-target.js'
+
 import NavigateMixin from '../../src/mixins/navigate-mixin.js';
 
 class HassioPagesWithTabs extends NavigateMixin(PolymerElement) {
@@ -41,10 +44,10 @@ class HassioPagesWithTabs extends NavigateMixin(PolymerElement) {
           </template>
         </app-toolbar>
         <paper-tabs scrollable="" selected="[[page]]" attr-for-selected="page-name" on-iron-activate="handlePageSelected">
-          <paper-tab on-click="_scrollToTop" page-name="dashboard">Dashboard</paper-tab>
-          <paper-tab on-click="_scrollToTop" page-name="snapshots">Snapshots</paper-tab>
-          <paper-tab on-click="_scrollToTop" page-name="store">Add-on store</paper-tab>
-          <paper-tab on-click="_scrollToTop" page-name="system">System</paper-tab>
+          <paper-tab page-name="dashboard">Dashboard</paper-tab>
+          <paper-tab page-name="snapshots">Snapshots</paper-tab>
+          <paper-tab page-name="store">Add-on store</paper-tab>
+          <paper-tab page-name="system">System</paper-tab>
         </paper-tabs>
       </app-header>
       <template is="dom-if" if="[[equals(page, &quot;dashboard&quot;)]]">
@@ -99,6 +102,7 @@ class HassioPagesWithTabs extends NavigateMixin(PolymerElement) {
     if (newPage !== this.page) {
       this.navigate(`/hassio/${newPage}`);
     }
+    scrollToTarget(this, this.$.layout.header.scrollTarget);
   }
 
   equals(a, b) {
@@ -123,45 +127,6 @@ class HassioPagesWithTabs extends NavigateMixin(PolymerElement) {
       markdownContent: ev.detail.content,
     });
     this.shadowRoot.querySelector('hassio-markdown-dialog').openDialog();
-  }
-
-  /**
-   * Scroll to a specific y coordinate.
-   *
-   * Copied from paper-scroll-header-panel.
-   *
-   * @method scroll
-   * @param {number} top The coordinate to scroll to, along the y-axis.
-   * @param {boolean} smooth true if the scroll position should be smoothly adjusted.
-   */
-  _scrollToTop() {
-    // the scroll event will trigger _updateScrollState directly,
-    // However, _updateScrollState relies on the previous `scrollTop` to update the states.
-    // Calling _updateScrollState will ensure that the states are synced correctly.
-    var top = 0;
-    var scroller = this.$.layout.header.scrollTarget;
-    var easingFn = function easeOutQuad(t, b, c, d) {
-      /* eslint-disable no-param-reassign, space-infix-ops, no-mixed-operators */
-      t /= d;
-      return -c * t*(t-2) + b;
-      /* eslint-enable no-param-reassign, space-infix-ops, no-mixed-operators */
-    };
-    var animationId = Math.random();
-    var duration = 200;
-    var startTime = Date.now();
-    var currentScrollTop = scroller.scrollTop;
-    var deltaScrollTop = top - currentScrollTop;
-    this._currentAnimationId = animationId;
-    (function updateFrame() {
-      var now = Date.now();
-      var elapsedTime = now - startTime;
-      if (elapsedTime > duration) {
-        scroller.scrollTop = top;
-      } else if (this._currentAnimationId === animationId) {
-        scroller.scrollTop = easingFn(elapsedTime, currentScrollTop, deltaScrollTop, duration);
-        requestAnimationFrame(updateFrame.bind(this));
-      }
-    }).call(this);
   }
 }
